Refactor structure of files
